refactor: split path geometry and rendering
